Split parsers/generic.hpp
- Move parse lines into lines.hpp
- Move to_number into number.hpp
- Move to_sequence into proc_stat.cpp
