+++ src/block.cpp
@@ -19,7 +19,7 @@
 #include "pfs/defer.hpp"
 #include "pfs/parsers/block_stat.hpp"
 #include "pfs/parsers/common.hpp"
-#include "pfs/parsers/generic.hpp"
+#include "pfs/parsers/number.hpp"
 #include "pfs/block.hpp"
 #include "pfs/utils.hpp"
