+++ src/parsers/proc_stat.cpp
@@ -17,7 +17,7 @@
 #include <chrono>
 #include <cstring>
 
-#include "pfs/parsers/generic.hpp"
+#include "pfs/parsers/number.hpp"
 #include "pfs/parsers/proc_stat.hpp"
 #include "pfs/utils.hpp"
 
@@ -26,6 +26,50 @@ namespace impl {
 namespace parsers {
 
 namespace {
+
+template <typename T>
+static void to_sequence(const std::string& value, proc_stat::sequence<T>& out)
+{
+    // Some examples:
+    // clang-format off
+    // 975101428 40707218 345522235 433770 2054357 19668 0 1807723 381659448 33954 202863055
+    // clang-format on
+
+    enum token
+    {
+        TOTAL     = 0,
+        MIN_COUNT = 1,
+    };
+
+    auto tokens = utils::split(value);
+    if (tokens.size() < MIN_COUNT)
+    {
+        throw parser_error("Corrupted sequence - Unexpected tokens count",
+                           value);
+    }
+
+    try
+    {
+        proc_stat::sequence<T> sequence;
+
+        utils::stot(tokens[TOTAL], out.total);
+
+        for (size_t i = MIN_COUNT; i < tokens.size(); i++)
+        {
+            unsigned long long value;
+            utils::stot(tokens[i], value);
+            out.per_item.push_back(value);
+        }
+    }
+    catch (const std::invalid_argument& ex)
+    {
+        throw parser_error("Corrupted sequence - Invalid argument", value);
+    }
+    catch (const std::out_of_range& ex)
+    {
+        throw parser_error("Corrupted sequence - Out of range", value);
+    }
+}
 
 static void to_cpu(const std::string& value, proc_stat::cpu& out)
 {
