+++ src/task.cpp
@@ -31,7 +31,7 @@
 #include "pfs/parsers/cgroup.hpp"
 #include "pfs/parsers/maps.hpp"
 #include "pfs/parsers/mountinfo.hpp"
-#include "pfs/parsers/generic.hpp"
+#include "pfs/parsers/lines.hpp"
 #include "pfs/parsers/common.hpp"
 #include "pfs/parsers/task_io.hpp"
 #include "pfs/parsers/task_status.hpp"
